requirements.txt
django==3.2.20
djangorestframework==3.14.0
django-environ==0.10.0
django-filter==23.2

psycopg2-binary==2.9.6

django-redis==5.3.0
redis==4.6.0
celery==5.3.1
django-celery-results==2.5.1

drf-yasg==1.21.6
coreapi==2.3.3
pillow==10.0.0
geopy==2.3.0

isort==5.12.0
flake8==6.0.0
